Img doc printing: Implement *ugly* workaround.
When printing an image, we do a PIL.open() + image2surface()
+ cairo.set_source_surface(). However, for some reason, it seems
the reference to the surface is lost too soon (ie before the
whole print is rendered). So the surface is garbage collected
too soon and Paperwork crashes when cairo tries to render the surface
on the context.

It's most likely a cairo + gobject introspection + python bug.

The workaround here is simply to keep the surface in a dictionary
as long as required.

Signed-off-by: Jerome Flesch <jflesch@gmail.com>