Fix search dialog with Python3
The search dialog did not work under python3 as the dialogs are returning
unicode "str" and the manual "encode()" and "decode()" that we were
performing failed. This made the cancel/OK buttons not work, among many
things :-). Tested with a range of search terms and keywords, and dates
and labels..