Fix crash on index rescan if unfortunately named files are present
I discovered that when syncing the paperwork workingdirectory between two pcs via syncthing (the "smartness" of that move is not under discussion here) that it wouldn't read the workingdirectory anymore..
After a bit of digging I found, that syncthing had created files like: paper.1.thumb.sync-conflict-20171019-151444-K454DB6.jpg in the doc-subdirs.

This file name made the code assume a wrong number of pages for this document and subsequently crash in add_doc.

This makes the parsing more stringent and thus fixing this crash. I tested it shortly with my workingdirectory and it seems to work nicely.

Signed-off-by: Florian Mickler <florian@mickler.org>
Cleaned-up-by: Jerome Flesch <jflesch@openpaper.work>